requirements.txt
progressbar2==3.37.1
opencv-python==4.1.0.25
numpy==1.17.2